backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claim-optimizer-backend"
version = "0.1.0"
description = "AI-assisted Insurance Claim Optimizer backend (FastAPI)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]",
  "pydantic>=2",
  "pydantic-settings",
  "orjson",
  "httpx",
  "python-jose[cryptography]",
  "sse-starlette",
]

[tool.ruff]
line-length = 120

